JVMCBC-507: Make sure core id is random and long.
Motivation
----------
Previously the core id would start at 0 all the time (1 as the
first core), but this makes it not unique across multiple
application servers.

Also, it should actually be stored as a long since its expected
to be dumped as 16 char hex value.

Modifications
-------------
To increase the entropy of the core id, it is initialized at
a random integer. Note that if its below 0, we force it positive
so that its also in bounds.

Result
------
When many app servers are used, its now easier to distinguish
them in the logs and via tracing.

Reviewed-on: http://review.couchbase.org/90131
Tested-by: Michael Nitschinger <michael@nitschinger.at>

+++ src/main/java/com/couchbase/client/core/CoreContext.java
@@ -34,7 +34,7 @@ public class CoreContext {
 
     private final CoreEnvironment environment;
     private final RingBuffer<ResponseEvent> responseRingBuffer;
-    private final int coreId;
+    private final long coreId;
 
     /**
      * Creates a new {@link CoreContext} with no core id.
@@ -56,7 +56,7 @@ public class CoreContext {
      */
     public CoreContext(final CoreEnvironment environment,
         final RingBuffer<ResponseEvent> responseRingBuffer,
-        final int coreId) {
+        final long coreId) {
         this.environment = environment;
         this.responseRingBuffer = responseRingBuffer;
         this.coreId = coreId;
@@ -79,7 +79,7 @@ public class CoreContext {
     /**
      * The core it, 0 if not set.
      */
-    public int coreId() {
+    public long coreId() {
         return coreId;
     }
 }
